flutter 3 dependency upgrade

+++ lib/helpers/decimal_handler.dart
@@ -12,7 +12,7 @@ Decimal toDecimal(BigInt amount, int decimals) {
   Decimal baseUnit = Decimal.fromInt(10).pow(decimals);
 
   var d = Decimal.parse(amount.toString());
-  d = d / baseUnit;
+  d = (d / baseUnit) as Decimal;
 
   return d;
 }
